refactor(datagrid): components folder
