+++ src/renderer/src/features/DataGrid/components/Section.tsx
