+++ src/renderer/src/features/DataGrid/components/index.ts
@@ -1,0 +1,5 @@
+import { Cell, CellWrapper } from "./Cell";
+import { Row } from "./Row";
+import { Section } from "./Section";
+
+export { Cell, CellWrapper, Row, Section };
